% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_simpleDE.R
\name{Create_simpleDE}
\alias{Create_simpleDE}
\title{Create_simpleDE}
\usage{
Create_simpleDE(expr_mtx, sample_info = NULL, gene_info = NULL)
}
\arguments{
\item{expr_mtx}{expression matrix (n genes * p cells) for expression data}

\item{sample_info}{dataframe of p rows, storing sample information. If it is not specified, a data.frame is created with one
column of colnames of expr_mtx as sample id.}

\item{gene_info}{dataframe of n rows, storing gene annotation information. If it is not specified, a data.frame is created with one
column of rownames of expr_mtx as gene id.}
}
\value{
a list containing all data and information
}
\description{
Creates a list containg expression matrix, sample information and gene annotation.
}
\examples{
# Create with only an expression matrix
a = matrix(rpois(12, 1), 4, 3)
rownames(a) = sapply(1:4, function(x) paste0("gene", x))
colnames(a) = sapply(1:3, function(x) paste0("sample",x))
simple.a = Create_simpleDE(a)

# Create with information of samples and genes
g = data.frame(ENTREZ_ID = c(123,124,125,126))
rownames(g) = sapply(1:4, function(x) paste0("gene", x))
s = data.frame(batch = c(1,2,2))
rownames(s) = sapply(1:3, function(x) paste0("sample",x))
simple.a = Create_simpleDE(a, s, g)


}
