% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA_test.R
\name{DEA_test}
\alias{DEA_test}
\title{DEA_test}
\usage{
DEA_test(x, group)
}
\arguments{
\item{x}{a list from Normalize_data}

\item{group}{a colname from sample_info or a vetor of factor indexing two group for comparison}

\item{verbose}{Print the information of reference and process}
}
\value{
a matrix containg fold change, test statistics and p value with adjustment
}
\description{
Perform differential expression analysis using t-test
}
\examples{
# Using colname from sample_info
de.a = DEA_test(simple.a, group = "batch")

# Using customized index
de.a = DEA_test(simple.a, group = c(1,1,2,2))

}
