% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter_Data.R
\name{Filter_Data}
\alias{Filter_Data}
\title{Filter_Data}
\usage{
Filter_Data(x, min_gene = 5, min_read = 5, verbose = T)
}
\arguments{
\item{x}{a list created from Create_simpleDE}

\item{min_gene}{filter out samples with expressed gene less than this}

\item{min_read}{filter out genes with total reads less than this}

\item{verbose}{Print the information of filtering}
}
\value{
the origin list with some genes and cells removed
}
\description{
Filter the genes with low expression and cells with low reads
}
\examples{
simple.a = Filter_Data(simple.a)

}
