% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_DEG.R
\name{Get_DEG}
\alias{Get_DEG}
\title{Get_DEG}
\usage{
Get_DEG(x, criteria = c("p.value", "log_FC", "fdr"), threshold, is.sort = T)
}
\arguments{
\item{x}{a data.frame containing results from DEA_test}

\item{criteria}{criteria to filter DEG, with options of p.value, absolute value of log_FC and fdr}

\item{threshold}{threshold of the criteria to filter DEG}

\item{is.sort}{whether sorting genes based on criteria}
}
\value{
a data.frame containing interesting gene
}
\description{
Get differentail expression genes based on optional criteria
}
\examples{
res1 = Get_DEG(de.a, criteria = "log_FC", threshold = 0.1)
# Multiple filtering
Get_DEG(res2, criteria = "fdr", threshold = 0.1)

}
