% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalize_Data.R
\name{Normalize_Data}
\alias{Normalize_Data}
\title{Normalize_Data}
\usage{
Normalize_Data(x, option = c("size_factor", "simplest", "median"))
}
\arguments{
\item{x}{a list from Filter_Data}

\item{option}{option of normalization:
"simplest": divided by mean reads of each cell
"median": divided by median reads of each cell
"size_factor": (default)divided by size factor, which is the median of the ratios of raw data and geometric mean
"none": not performing normalization}
}
\value{
a list adding with a new element containing normalized data
}
\description{
Normalize raw data with optional normalization techniques
}
\examples{
a = matrix(c(1:8,1:8), 4, 4)
rownames(a) = sapply(1:4, function(x) paste0("gene", x))
colnames(a) = sapply(1:4, function(x) paste0("sample",x))
g = data.frame(ENTREZ_ID = c(123,124,125,126))
rownames(g) = sapply(1:4, function(x) paste0("gene", x))
s = data.frame(batch = c(1,2,2,1))
rownames(s) = sapply(1:4, function(x) paste0("sample",x))
simple.a = Create_simpleDE(a, s, g)
simple.a = Filter_Data(simple.a, 0, 0)
simple.a = Normalize_Data(simple.a, "size_factor")

}
