% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEA_test.R
\name{DEA_test}
\alias{DEA_test}
\title{DEA_test}
\usage{
DEA_test(x, group)
}
\arguments{
\item{x}{a list from Normalize_data}

\item{group}{a colname from sample_info or a vetor of factor indexing two group for comparison}

\item{verbose}{Print the information of reference and process}
}
\value{
a matrix containg fold change, test statistics and p value with adjustment
}
\description{
Perform differential expression analysis using t-test
}
\examples{
a = matrix(c(1:8,1:8), 4, 4)
rownames(a) = sapply(1:4, function(x) paste0("gene", x))
colnames(a) = sapply(1:4, function(x) paste0("sample",x))
g = data.frame(ENTREZ_ID = c(123,124,125,126))
rownames(g) = sapply(1:4, function(x) paste0("gene", x))
s = data.frame(batch = c(1,2,2,1))
rownames(s) = sapply(1:4, function(x) paste0("sample",x))
simple.a = Create_simpleDE(a, s, g)
simple.a = Filter_Data(simple.a, 0, 0)
simple.a = Normalize_Data(simple.a, "size_factor")

# Using colname from sample_info
de.a = DEA_test(simple.a, group = "batch")

# Using customized index
de.a = DEA_test(simple.a, group = c(1,1,2,2))

}
