% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter_Data.R
\name{Filter_Data}
\alias{Filter_Data}
\title{Filter_Data}
\usage{
Filter_Data(x, min_gene = 5, min_read = 5, verbose = T)
}
\arguments{
\item{x}{a list created from Create_simpleDE}

\item{min_gene}{filter out samples with expressed gene less than this}

\item{min_read}{filter out genes with total reads less than this}

\item{verbose}{Print the information of filtering}
}
\value{
the origin list with some genes and cells removed
}
\description{
Filter the genes with low expression and cells with low reads
}
\examples{
a = matrix(c(1:8,1:8), 4, 4)
rownames(a) = sapply(1:4, function(x) paste0("gene", x))
colnames(a) = sapply(1:4, function(x) paste0("sample",x))
g = data.frame(ENTREZ_ID = c(123,124,125,126))
rownames(g) = sapply(1:4, function(x) paste0("gene", x))
s = data.frame(batch = c(1,2,2,1))
rownames(s) = sapply(1:4, function(x) paste0("sample",x))
simple.a = Create_simpleDE(a, s, g)
simple.a = Filter_Data(simple.a, 0, 0)

}
