% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalize_Data.R
\name{Normalize_Data}
\alias{Normalize_Data}
\title{Normalize_Data}
\usage{
Normalize_Data(x, option = c("size_factor", "simplest", "median"))
}
\arguments{
\item{x}{a list from Filter_Data}

\item{option}{option of normalization:
"simplest": divided by mean reads of each cell
"median": divided by median reads of each cell
"size_factor": (default)divided by size factor, which is the median of the ratios of raw data and geometric mean
"none": not performing normalization}
}
\value{
a list adding with a new element containing normalized data
}
\description{
Normalize raw data with optional normalization techniques
}
\examples{
simple.a = Normalize_Data(simple.a, option = "simplest")

}
